Move functions into proper repositories
