+++ Kongres.Api/Kongres.Api.Infrastructure/Repositories/Interfaces/IReviewRepository.cs
@@ -5,7 +5,6 @@ namespace Kongres.Api.Infrastructure.Repositories.Interfaces
 {
     public interface IReviewRepository : IRepository
     {
-        Task<bool> IsReviewerAsync(uint scientificWorkId, uint userId);
         Task<Review> GetReviewByIdAsync(uint reviewId);
         Task AddAnswerToReviewAsync(Review review);
         Task AddReviewAsync(Review review);
