+++ Kongres.Api/Kongres.Api.Infrastructure/Repositories/Interfaces/IScientificWorkRepository.cs
@@ -11,7 +11,6 @@ namespace Kongres.Api.Infrastructure.Repositories.Interfaces
         Task<IEnumerable<ScientificWork>> GetApprovedWorksAsync();
         Task<ScientificWork> GetWorkByIdAsync(uint scientificWorkId);
         Task<bool> IsAuthorOfScientificWorkByReviewIdAsync(uint userId, uint reviewOfWorkId);
-        Task<bool> IsReviewerOfScientificWorkAsync(uint userId, uint scientificWorkId);
         Task<byte> GetNumberOfVersionsByAuthorIdAsync(uint userId);
         Task<IEnumerable<ScientificWork>> GetAllBySpecializationAsync(string specialization);
         Task ChangeStatusAsync(ScientificWork scientificWork);
