+++ Kongres.Api/Kongres.Api.Infrastructure/Repositories/ReviewerScientificWorkRepository.cs
@@ -37,5 +37,19 @@ namespace Kongres.Api.Infrastructure.Repositories
 
         public int GetReviewersCount(uint scientificWorkId)
             => _context.ReviewersScienceWorks.Count(x => x.ScientificWork.Id == scientificWorkId);
+
+
+        // check if reviewer is assign to given scientific work
+        public async Task<bool> IsReviewerAsync(uint scientificWorkId, uint userId)
+            => await _context.ReviewersScienceWorks.Include(x => x.User)
+                                                   .Include(x => x.ScientificWork)
+                                                   .AnyAsync(x => x.ScientificWork.Id == scientificWorkId &&
+                                                                  x.User.Id == userId);
+
+        public async Task<bool> IsReviewerOfScientificWorkAsync(uint userId, uint scientificWorkId)
+            => await _context.ReviewersScienceWorks.Include(x => x.ScientificWork)
+                                                   .Include(x => x.User)
+                                                   .AnyAsync(x => x.User.Id == userId &&
+                                                                  x.ScientificWork.Id == scientificWorkId);
     }
 }
