+++ Kongres.Api/Kongres.Api.Infrastructure/Repositories/UserRepository.cs
@@ -1,5 +1,10 @@
-﻿using Kongres.Api.Infrastructure.Context;
+﻿using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using Kongres.Api.Domain.Entities;
+using Kongres.Api.Infrastructure.Context;
 using Kongres.Api.Infrastructure.Repositories.Interfaces;
+using Microsoft.EntityFrameworkCore;
 
 namespace Kongres.Api.Infrastructure.Repositories
 {
@@ -14,5 +19,10 @@ namespace Kongres.Api.Infrastructure.Repositories
 
         public string GetEmailById(uint userId)
             => _context.Users.Find(userId).Email;
+
+        public async Task<IEnumerable<User>> GetAllBySpecializationAsync(string specialization)
+            => await _context.Users.Where(x => x.Specialization == specialization &&
+                                               x.NormalizedUserName.Contains("REVIEWER"))
+                                   .ToListAsync();
     }
 }
